requirements.txt
entrypoints==0.3
flake8==3.7.8
mccabe==0.6.1
mysqlclient==1.4.4
pycodestyle==2.5.0
pyflakes==2.1.1
python-dotenv==0.10.3
SQLAlchemy==1.3.10
